requirements.txt
Flask
pandas
numpy
scikit-learn
faiss-cpu
